Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
mysql-connector-python==8.1.0
cryptography==41.0.3
pymysql==1.1.0
pandas
openpyxl
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
